requirements.txt
django==3.1.3
django-admin
django-filter
djangorestframework==3.12.4
psycopg2-binary==2.8.6
django-localflavor==3.0.1